did the last branch, sorta
I need to rewrite the ids for each choice and had like 600 to it. I also need to work on the combat, since I made new stories I need to create new versions of combat for each story. I also need to link the parts that are commented to fully complete the story part. After I will work on the imaging and other things like that, but we will see how much time we have left and how much energy we have.