Welp I guess this goes my grade
We tried our hardest to push to the due date, we have spent the last week doing code for 10 hours, we are only able to do this. We still need to reorganized the text, we thought our way would work but it sadly didn't. We also have enough time to make images now. We have one more chance, one more hope,  but I do not think we will be able to complete it within 30 minutes. My vision is blurry this has to be one of my most worked on projects, sucks that we will not get the credit we derive. Well there will be another commit, but this is last comments I will have. Good night, sorry that we can not finish in time, please of mercy, we tried our hardest.