We are done
we will see if things go our way but we completed the imaging and bug fixed for the past few hours. Hopefully we have tomorrow to work on it too, so we can do some cool things. Thankyou for everything you do, remeber someone cares and as always have a good day!